requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
validators>=0.20.0
ratelimit>=2.2.1 